Split up decision making components into GameEngine and tests
